rename pandas jobs
